+++ game/theater/start_generator.py
@@ -85,6 +85,7 @@ class ModSettings:
     fa_18efg: bool = False
     fa18ef_tanker: bool = False
     f22_raptor: bool = False
+    f23a_black_widow: bool = False
     f84g_thunderjet: bool = False
     f100_supersabre: bool = False
     f104_starfighter: bool = False